Fix test mock for commits() to return iterator
The test_username_lookup_case_insensitive test was failing because it mocked
repo_data.commits() to return a list, but the main function expects an iterator
that can be used with next(). Also added explicit owning_team=None to ensure
the test takes the original commit author code path for case-insensitive lookup.

Fixes TypeError: list object is not an iterator

Signed-off-by: Zack Koppert <zkoppert@github.com>